requirements.txt
# Core scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
urllib3==2.0.7

# Web framework (using Flask for simplicity)
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0

# Database and Authentication
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
psycopg2-binary==2.9.9
email-validator==2.1.0

# Additional utilities
python-dotenv==1.0.1
APScheduler==3.10.4
Werkzeug==3.0.1